+++ plugins/app/src/components/teams/Team.tsx
@@ -10,9 +10,9 @@ import {
   TextContent,
 } from '@patternfly/react-core';
 import { QueryObserverResult, useQuery } from 'react-query';
-import React, { useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { ExternalLinkAltIcon } from '@patternfly/react-icons';
+import React from 'react';
 
 import { PageContentSection, PageHeaderSection } from '@kobsio/shared';
 import { DashboardsWrapper } from '../dashboards/DashboardsWrapper';
@@ -25,7 +25,6 @@ interface ITeamParams extends Record<string, string | undefined> {
 const Team: React.FunctionComponent = () => {
   const navigate = useNavigate();
   const params = useParams<ITeamParams>();
-  const [details, setDetails] = useState<React.ReactNode>(undefined);
 
   const { isError, isLoading, error, data, refetch } = useQuery<ITeam, Error>(
     ['app/teams/team', params.team],
@@ -108,9 +107,14 @@ const Team: React.FunctionComponent = () => {
         }
       />
 
-      <PageContentSection hasPadding={false} hasDivider={false} toolbarContent={undefined} panelContent={details}>
+      <PageContentSection
+        hasPadding={false}
+        hasDivider={data.dashboards ? false : true}
+        toolbarContent={undefined}
+        panelContent={undefined}
+      >
         {data.dashboards ? (
-          <DashboardsWrapper manifest={data} references={data.dashboards} setDetails={setDetails} />
+          <DashboardsWrapper manifest={data} references={data.dashboards} useDrawer={true} />
         ) : (
           <div></div>
         )}