+++ plugins/plugin-jaeger/src/components/page/Traces.tsx
@@ -25,9 +25,11 @@ const Traces: React.FunctionComponent<ITracesProps> = ({ instance }: ITracesProp
     navigate(
       `${location.pathname}?limit=${opts.limit}&maxDuration=${opts.maxDuration}&minDuration=${
         opts.minDuration
-      }&operation=${opts.operation === 'All Operations' ? '' : opts.operation}&service=${opts.service}&tags=${
-        opts.tags
-      }&time=${opts.times.time}&timeEnd=${opts.times.timeEnd}&timeStart=${opts.times.timeStart}`,
+      }&operation=${
+        opts.operation === 'All Operations' ? '' : encodeURIComponent(opts.operation)
+      }&service=${encodeURIComponent(opts.service)}&tags=${encodeURIComponent(opts.tags)}&time=${
+        opts.times.time
+      }&timeEnd=${opts.times.timeEnd}&timeStart=${opts.times.timeStart}`,
     );
   };